+++ pMuTT/__init__.py
@@ -10,7 +10,7 @@ pMuTT
 # present, too:
 #
 name = 'pMuTT'
-__version__ = '1.2.0'
+__version__ = '1.2.1'
 
 import re
 import inspect